Revert "LPS-42120 Rename"
This reverts commit 9f74b6577f02e033327a65340004d4987485da47.
