+++ portal-impl/src/com/liferay/portal/deploy/auto/ModuleAutoDeployListener.java
@@ -18,7 +18,6 @@ import aQute.bnd.header.OSGiHeader;
 import aQute.bnd.header.Parameters;
 import aQute.bnd.osgi.Constants;
 
-import com.liferay.portal.deploy.auto.util.PluginAutoDetectorUtil;
 import com.liferay.portal.kernel.deploy.auto.AutoDeployException;
 import com.liferay.portal.kernel.deploy.auto.AutoDeployer;
 import com.liferay.portal.kernel.deploy.auto.BaseAutoDeployListener;
@@ -62,8 +61,8 @@ public class ModuleAutoDeployListener extends BaseAutoDeployListener {
 	}
 
 	protected boolean isModule(File file) throws AutoDeployException {
-		PluginAutoDetectorUtil pluginAutoDeployListenerHelper =
-			new PluginAutoDetectorUtil(file);
+		PluginAutoDeployListenerHelper pluginAutoDeployListenerHelper =
+			new PluginAutoDeployListenerHelper(file);
 
 		if (!pluginAutoDeployListenerHelper.isJarFile()) {
 			return false;
